move config inside logger
