--- config/BUILD.bazel
@@ -1,12 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_test(
-    name = "go_default_test",
-    embed = [":go_default_library"],
-    srcs = [
-        "config_test.go",
-    ],
-    deps =[
-        "@com_github_stretchr_testify//assert:go_default_library"
-    ]
-)
